model: move event to gateway, add deserialization
Move the Event enum and its related types (such as the `EventType`,
`ShardEvent` and child structs, `DispatchEvent`, and `GatewayEvent`)
into the `model::gateway` module in a new `event` submodule.

Additionally, add a `DispatchEventWithTypeDeserializer` seeded
deserializer. The deserializer is given a seed of the type name received
by the gateway, which can be used for efficient deserialization.

Closes #102.
