+++ gateway/examples/cluster/src/main.rs
@@ -21,7 +21,7 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync>> {
     let mut events = cluster.events().await;
 
     while let Some((id, event)) = events.next().await {
-        println!("Shard: {}, Event: {:?}", id, event.event_type());
+        println!("Shard: {}, Event: {:?}", id, event.kind());
     }
 
     Ok(())
