+++ gateway/src/event.rs
@@ -1,410 +1,143 @@
-#![allow(clippy::wildcard_imports)]
-use serde::{
-    de::{Deserialize, Deserializer, Error as DeError, MapAccess, Visitor},
-    Deserialize as DeserializeMacro,
-};
-use serde_value::{DeserializerError as ValueDeserializerError, Value};
-use std::{
-    convert::TryFrom,
-    fmt::{Formatter, Result as FmtResult},
-};
-use twilight_model::gateway::{payload::*, OpCode};
+use bitflags::bitflags;
+use twilight_model::gateway::event::EventType;
 
-#[cfg(feature = "metrics")]
-use metrics::counter;
-
-/// An event from the gateway, which can either be a dispatch event with
-/// stateful updates or a heartbeat, hello, etc. that a shard needs to operate.
-#[derive(Clone, Debug)]
-pub enum GatewayEvent {
-    Dispatch(u64, Box<DispatchEvent>),
-    Heartbeat(u64),
-    HeartbeatAck,
-    Hello(u64),
-    InvalidateSession(bool),
-    Reconnect,
-}
-
-#[derive(DeserializeMacro)]
-#[serde(field_identifier, rename_all = "lowercase")]
-enum Field {
-    D,
-    Op,
-    S,
-    T,
-}
-
-struct GatewayEventVisitor;
-
-impl<'de> Visitor<'de> for GatewayEventVisitor {
-    type Value = GatewayEvent;
-
-    fn expecting(&self, formatter: &mut Formatter<'_>) -> FmtResult {
-        formatter.write_str("struct GatewayEvent")
+bitflags! {
+    /// Bitflags representing all of the possible types of events.
+    pub struct EventTypeFlags: u64 {
+        /// A user was banned from a guild.
+        const BAN_ADD = 1;
+        /// A user was unbanned from a guild.
+        const BAN_REMOVE = 1 << 1;
+        /// A channel was created.
+        const CHANNEL_CREATE = 1 << 2;
+        /// A channel was deleted.
+        const CHANNEL_DELETE = 1 << 3;
+        /// A channel's pins were updated.
+        const CHANNEL_PINS_UPDATE = 1 << 4;
+        /// A channel was updated.
+        const CHANNEL_UPDATE = 1 << 5;
+        /// A heartbeat was created.
+        const GATEWAY_HEARTBEAT = 1 << 6;
+        /// A heartbeat was acknowledged.
+        const GATEWAY_HEARTBEAT_ACK = 1 << 7;
+        /// A "hello" packet was received from the gateway.
+        const GATEWAY_HELLO = 1 << 8;
+        /// A shard's session was invalidated.
+        ///
+        /// `true` if resumeable. If not, then the shard must do a full
+        /// reconnect.
+        const GATEWAY_INVALIDATE_SESSION = 1 << 8;
+        /// The gateway is indicating to perform a reconnect.
+        const GATEWAY_RECONNECT = 1 << 9;
+        /// A guild was created.
+        const GUILD_CREATE = 1 << 10;
+        /// A guild was deleted or the current user was removed from a guild.
+        const GUILD_DELETE = 1 << 11;
+        /// A guild's emojis were updated.
+        const GUILD_EMOJIS_UPDATE = 1 << 12;
+        /// A guild's integrations were updated.
+        const GUILD_INTEGRATIONS_UPDATE = 1 << 13;
+        /// A guild was updated.
+        const GUILD_UPDATE = 1 << 14;
+        const INVITE_CREATE = 1 << 46;
+        const INVITE_DELETE = 1 << 47;
+        const MEMBER_ADD = 1 << 15;
+        const MEMBER_REMOVE = 1 << 16;
+        const MEMBER_UPDATE = 1 << 17;
+        const MEMBER_CHUNK = 1 << 18;
+        const MESSAGE_CREATE = 1 << 19;
+        const MESSAGE_DELETE = 1 << 20;
+        const MESSAGE_DELETE_BULK = 1 << 21;
+        const MESSAGE_UPDATE = 1 << 22;
+        const PRESENCE_UPDATE = 1 << 23;
+        const PRESENCES_REPLACE = 1 << 24;
+        const REACTION_ADD = 1 << 25;
+        const REACTION_REMOVE = 1 << 26;
+        const REACTION_REMOVE_ALL = 1 << 27;
+        const REACTION_REMOVE_EMOJI = 1 << 48;
+        const READY = 1 << 28;
+        const RESUMED = 1 << 29;
+        const ROLE_CREATE = 1 << 30;
+        const ROLE_DELETE = 1 << 31;
+        const ROLE_UPDATE = 1 << 32;
+        const SHARD_CONNECTED = 1 << 33;
+        const SHARD_CONNECTING = 1 << 34;
+        const SHARD_DISCONNECTED = 1 << 35;
+        const SHARD_IDENTIFYING = 1 << 36;
+        const SHARD_PAYLOAD = 1 << 45;
+        const SHARD_RECONNECTING = 1 << 37;
+        const SHARD_RESUMING = 1 << 38;
+        const TYPING_START = 1 << 39;
+        const UNAVAILABLE_GUILD = 1 << 40;
+        const USER_UPDATE = 1 << 41;
+        const VOICE_SERVER_UPDATE = 1 << 42;
+        const VOICE_STATE_UPDATE = 1 << 43;
+        const WEBHOOKS_UPDATE = 1 << 44;
     }
+}
 
-    fn visit_map<V>(self, mut map: V) -> Result<GatewayEvent, V::Error>
-    where
-        V: MapAccess<'de>,
-    {
-        static VALID_OPCODES: &[&str] = &[
-            "EVENT",
-            "HEARTBEAT",
-            "HEARTBEAT_ACK",
-            "HELLO",
-            "IDENTIFY",
-            "INVALID_SESSION",
-            "RECONNECT",
-        ];
-
-        // Have to use a serde_json::Value here because serde has no
-        // abstract container type.
-        let mut d = None::<Value>;
-        let mut op = None::<OpCode>;
-        let mut s = None::<u64>;
-        let mut t = None::<String>;
-
-        while let Some(key) = map.next_key()? {
-            match key {
-                Field::D => {
-                    if d.is_some() {
-                        return Err(DeError::duplicate_field("d"));
-                    }
-
-                    d = Some(map.next_value()?);
-                }
-                Field::Op => {
-                    if op.is_some() {
-                        return Err(DeError::duplicate_field("op"));
-                    }
-
-                    op = Some(map.next_value()?);
-                }
-                Field::S => {
-                    if s.is_some() {
-                        return Err(DeError::duplicate_field("s"));
-                    }
-
-                    s = map.next_value::<Option<_>>()?;
-                }
-                Field::T => {
-                    if t.is_some() {
-                        return Err(DeError::duplicate_field("t"));
-                    }
-
-                    t = map.next_value::<Option<_>>()?;
-                }
-            }
+impl From<EventType> for EventTypeFlags {
+    fn from(event_type: EventType) -> Self {
+        match event_type {
+            EventType::BanAdd => EventTypeFlags::BAN_ADD,
+            EventType::BanRemove => EventTypeFlags::BAN_REMOVE,
+            EventType::ChannelCreate => EventTypeFlags::CHANNEL_CREATE,
+            EventType::ChannelDelete => EventTypeFlags::CHANNEL_DELETE,
+            EventType::ChannelPinsUpdate => EventTypeFlags::CHANNEL_PINS_UPDATE,
+            EventType::ChannelUpdate => EventTypeFlags::CHANNEL_UPDATE,
+            EventType::GatewayHeartbeat => EventTypeFlags::GATEWAY_HEARTBEAT,
+            EventType::GatewayHeartbeatAck => EventTypeFlags::GATEWAY_HEARTBEAT_ACK,
+            EventType::GatewayHello => EventTypeFlags::GATEWAY_HELLO,
+            EventType::GatewayInvalidateSession => EventTypeFlags::GATEWAY_INVALIDATE_SESSION,
+            EventType::GatewayReconnect => EventTypeFlags::GATEWAY_RECONNECT,
+            EventType::GuildCreate => EventTypeFlags::GUILD_CREATE,
+            EventType::GuildDelete => EventTypeFlags::GUILD_DELETE,
+            EventType::GuildEmojisUpdate => EventTypeFlags::GUILD_EMOJIS_UPDATE,
+            EventType::GuildIntegrationsUpdate => EventTypeFlags::GUILD_INTEGRATIONS_UPDATE,
+            EventType::GuildUpdate => EventTypeFlags::GUILD_UPDATE,
+            EventType::InviteCreate => EventTypeFlags::INVITE_CREATE,
+            EventType::InviteDelete => EventTypeFlags::INVITE_DELETE,
+            EventType::MemberAdd => EventTypeFlags::MEMBER_ADD,
+            EventType::MemberRemove => EventTypeFlags::MEMBER_REMOVE,
+            EventType::MemberUpdate => EventTypeFlags::MEMBER_UPDATE,
+            EventType::MemberChunk => EventTypeFlags::MEMBER_CHUNK,
+            EventType::MessageCreate => EventTypeFlags::MESSAGE_CREATE,
+            EventType::MessageDelete => EventTypeFlags::MESSAGE_DELETE,
+            EventType::MessageDeleteBulk => EventTypeFlags::MESSAGE_DELETE_BULK,
+            EventType::MessageUpdate => EventTypeFlags::MESSAGE_UPDATE,
+            EventType::PresenceUpdate => EventTypeFlags::PRESENCE_UPDATE,
+            EventType::PresencesReplace => EventTypeFlags::PRESENCES_REPLACE,
+            EventType::ReactionAdd => EventTypeFlags::REACTION_ADD,
+            EventType::ReactionRemove => EventTypeFlags::REACTION_REMOVE,
+            EventType::ReactionRemoveAll => EventTypeFlags::REACTION_REMOVE_ALL,
+            EventType::ReactionRemoveEmoji => EventTypeFlags::REACTION_REMOVE_EMOJI,
+            EventType::Ready => EventTypeFlags::READY,
+            EventType::Resumed => EventTypeFlags::RESUMED,
+            EventType::RoleCreate => EventTypeFlags::ROLE_CREATE,
+            EventType::RoleDelete => EventTypeFlags::ROLE_DELETE,
+            EventType::RoleUpdate => EventTypeFlags::ROLE_UPDATE,
+            EventType::ShardConnected => EventTypeFlags::SHARD_CONNECTED,
+            EventType::ShardConnecting => EventTypeFlags::SHARD_CONNECTING,
+            EventType::ShardDisconnected => EventTypeFlags::SHARD_DISCONNECTED,
+            EventType::ShardIdentifying => EventTypeFlags::SHARD_IDENTIFYING,
+            EventType::ShardReconnecting => EventTypeFlags::SHARD_RECONNECTING,
+            EventType::ShardPayload => EventTypeFlags::SHARD_PAYLOAD,
+            EventType::ShardResuming => EventTypeFlags::SHARD_RESUMING,
+            EventType::TypingStart => EventTypeFlags::TYPING_START,
+            EventType::UnavailableGuild => EventTypeFlags::UNAVAILABLE_GUILD,
+            EventType::UserUpdate => EventTypeFlags::USER_UPDATE,
+            EventType::VoiceServerUpdate => EventTypeFlags::VOICE_SERVER_UPDATE,
+            EventType::VoiceStateUpdate => EventTypeFlags::VOICE_STATE_UPDATE,
+            EventType::WebhooksUpdate => EventTypeFlags::WEBHOOKS_UPDATE,
         }
-
-        let op = op.ok_or_else(|| DeError::missing_field("op"))?;
-
-        Ok(match op {
-            OpCode::Event => {
-                let d = d.ok_or_else(|| DeError::missing_field("d"))?;
-                let s = s.ok_or_else(|| DeError::missing_field("s"))?;
-                let t = t.ok_or_else(|| DeError::missing_field("t"))?;
-                let dispatch = DispatchEvent::try_from((t.as_ref(), d)).map_err(DeError::custom)?;
-                #[cfg(feature = "metrics")]
-                counter!("DispatchEvent", 1, "DispatchEvent" => t);
-                GatewayEvent::Dispatch(s, Box::new(dispatch))
-            }
-            OpCode::Heartbeat => {
-                let s = s.ok_or_else(|| DeError::missing_field("s"))?;
-
-                GatewayEvent::Heartbeat(s)
-            }
-            OpCode::HeartbeatAck => GatewayEvent::HeartbeatAck,
-            OpCode::Hello => {
-                #[derive(DeserializeMacro)]
-                struct Hello {
-                    heartbeat_interval: u64,
-                }
-
-                let d = d.ok_or_else(|| DeError::missing_field("d"))?;
-                let hello = Hello::deserialize(d).map_err(DeError::custom)?;
-
-                GatewayEvent::Hello(hello.heartbeat_interval)
-            }
-            OpCode::InvalidSession => {
-                let d = d.ok_or_else(|| DeError::missing_field("d"))?;
-                let resumeable = bool::deserialize(d).map_err(DeError::custom)?;
-
-                GatewayEvent::InvalidateSession(resumeable)
-            }
-            OpCode::Identify => return Err(DeError::unknown_variant("Identify", VALID_OPCODES)),
-            OpCode::Reconnect => GatewayEvent::Reconnect,
-            OpCode::RequestGuildMembers => {
-                return Err(DeError::unknown_variant(
-                    "RequestGuildMembers",
-                    VALID_OPCODES,
-                ))
-            }
-            OpCode::Resume => return Err(DeError::unknown_variant("Resume", VALID_OPCODES)),
-            OpCode::StatusUpdate => {
-                return Err(DeError::unknown_variant("StatusUpdate", VALID_OPCODES))
-            }
-            OpCode::VoiceServerPing => {
-                return Err(DeError::unknown_variant("VoiceServerPing", VALID_OPCODES))
-            }
-            OpCode::VoiceStateUpdate => {
-                return Err(DeError::unknown_variant("VoiceStateUpdate", VALID_OPCODES))
-            }
-        })
     }
 }
 
-impl<'de> Deserialize<'de> for GatewayEvent {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        const FIELDS: &[&str] = &["d", "op", "s", "t"];
+impl Default for EventTypeFlags {
+    fn default() -> Self {
+        let mut flags = Self::all();
+        flags.remove(Self::SHARD_PAYLOAD);
 
-        deserializer.deserialize_struct("GatewayEvent", FIELDS, GatewayEventVisitor)
-    }
-}
-
-/// A dispatch event, containing information about a created guild, a member
-/// added, etc.
-#[derive(Clone, Debug)]
-pub enum DispatchEvent {
-    BanAdd(BanAdd),
-    BanRemove(BanRemove),
-    ChannelCreate(ChannelCreate),
-    ChannelDelete(ChannelDelete),
-    ChannelPinsUpdate(ChannelPinsUpdate),
-    ChannelUpdate(ChannelUpdate),
-    GuildCreate(Box<GuildCreate>),
-    GuildDelete(Box<GuildDelete>),
-    GuildEmojisUpdate(GuildEmojisUpdate),
-    GuildIntegrationsUpdate(GuildIntegrationsUpdate),
-    GuildUpdate(Box<GuildUpdate>),
-    InviteCreate(Box<InviteCreate>),
-    InviteDelete(InviteDelete),
-    MemberAdd(Box<MemberAdd>),
-    MemberRemove(MemberRemove),
-    MemberUpdate(Box<MemberUpdate>),
-    MemberChunk(MemberChunk),
-    MessageCreate(Box<MessageCreate>),
-    MessageDelete(MessageDelete),
-    MessageDeleteBulk(MessageDeleteBulk),
-    MessageUpdate(Box<MessageUpdate>),
-    PresenceUpdate(Box<PresenceUpdate>),
-    PresencesReplace,
-    ReactionAdd(Box<ReactionAdd>),
-    ReactionRemove(Box<ReactionRemove>),
-    ReactionRemoveAll(ReactionRemoveAll),
-    ReactionRemoveEmoji(ReactionRemoveEmoji),
-    Ready(Box<Ready>),
-    Resumed,
-    RoleCreate(RoleCreate),
-    RoleDelete(RoleDelete),
-    RoleUpdate(RoleUpdate),
-    TypingStart(Box<TypingStart>),
-    UnavailableGuild(UnavailableGuild),
-    UserUpdate(UserUpdate),
-    VoiceServerUpdate(VoiceServerUpdate),
-    VoiceStateUpdate(Box<VoiceStateUpdate>),
-    WebhooksUpdate(WebhooksUpdate),
-}
-
-impl TryFrom<(&str, Value)> for DispatchEvent {
-    type Error = ValueDeserializerError;
-
-    fn try_from((kind, v): (&str, Value)) -> Result<Self, Self::Error> {
-        Ok(match kind {
-            "CHANNEL_CREATE" => Self::ChannelCreate(ChannelCreate::deserialize(v)?),
-            "CHANNEL_DELETE" => Self::ChannelDelete(ChannelDelete::deserialize(v)?),
-            "CHANNEL_PINS_UPDATE" => Self::ChannelPinsUpdate(ChannelPinsUpdate::deserialize(v)?),
-            "CHANNEL_UPDATE" => Self::ChannelUpdate(ChannelUpdate::deserialize(v)?),
-            "GUILD_BAN_ADD" => Self::BanAdd(BanAdd::deserialize(v)?),
-            "GUILD_BAN_REMOVE" => Self::BanRemove(BanRemove::deserialize(v)?),
-            "GUILD_CREATE" => Self::GuildCreate(Box::new(GuildCreate::deserialize(v)?)),
-            "GUILD_DELETE" => Self::GuildDelete(Box::new(GuildDelete::deserialize(v)?)),
-            "GUILD_EMOJIS_UPDATE" => Self::GuildEmojisUpdate(GuildEmojisUpdate::deserialize(v)?),
-            "GUILD_INTEGRATIONS_UPDATE" => {
-                Self::GuildIntegrationsUpdate(GuildIntegrationsUpdate::deserialize(v)?)
-            }
-            "GUILD_MEMBERS_CHUNK" => Self::MemberChunk(MemberChunk::deserialize(v)?),
-            "GUILD_MEMBER_ADD" => Self::MemberAdd(Box::new(MemberAdd::deserialize(v)?)),
-            "GUILD_MEMBER_REMOVE" => Self::MemberRemove(MemberRemove::deserialize(v)?),
-            "GUILD_MEMBER_UPDATE" => Self::MemberUpdate(Box::new(MemberUpdate::deserialize(v)?)),
-            "GUILD_ROLE_CREATE" => Self::RoleCreate(RoleCreate::deserialize(v)?),
-            "GUILD_ROLE_DELETE" => Self::RoleDelete(RoleDelete::deserialize(v)?),
-            "GUILD_ROLE_UPDATE" => Self::RoleUpdate(RoleUpdate::deserialize(v)?),
-            "GUILD_UPDATE" => Self::GuildUpdate(Box::new(GuildUpdate::deserialize(v)?)),
-            "INVITE_CREATE" => Self::InviteCreate(Box::new(InviteCreate::deserialize(v)?)),
-            "INVITE_DELETE" => Self::InviteDelete(InviteDelete::deserialize(v)?),
-            "MESSAGE_CREATE" => Self::MessageCreate(Box::new(MessageCreate::deserialize(v)?)),
-            "MESSAGE_DELETE" => Self::MessageDelete(MessageDelete::deserialize(v)?),
-            "MESSAGE_DELETE_BULK" => Self::MessageDeleteBulk(MessageDeleteBulk::deserialize(v)?),
-            "MESSAGE_REACTION_ADD" => Self::ReactionAdd(Box::new(ReactionAdd::deserialize(v)?)),
-            "MESSAGE_REACTION_REMOVE" => {
-                Self::ReactionRemove(Box::new(ReactionRemove::deserialize(v)?))
-            }
-            "MESSAGE_REACTION_REMOVE_EMOJI" => {
-                Self::ReactionRemoveEmoji(ReactionRemoveEmoji::deserialize(v)?)
-            }
-            "MESSAGE_REACTION_REMOVE_ALL" => {
-                Self::ReactionRemoveAll(ReactionRemoveAll::deserialize(v)?)
-            }
-            "MESSAGE_UPDATE" => Self::MessageUpdate(Box::new(MessageUpdate::deserialize(v)?)),
-            "PRESENCE_UPDATE" => Self::PresenceUpdate(Box::new(PresenceUpdate::deserialize(v)?)),
-            "PRESENCES_REPLACE" => Self::PresencesReplace,
-            "READY" => Self::Ready(Box::new(Ready::deserialize(v)?)),
-            "RESUMED" => Self::Resumed,
-            "TYPING_START" => Self::TypingStart(Box::new(TypingStart::deserialize(v)?)),
-            "USER_UPDATE" => Self::UserUpdate(UserUpdate::deserialize(v)?),
-            "VOICE_SERVER_UPDATE" => Self::VoiceServerUpdate(VoiceServerUpdate::deserialize(v)?),
-            "VOICE_STATE_UPDATE" => {
-                Self::VoiceStateUpdate(Box::new(VoiceStateUpdate::deserialize(v)?))
-            }
-            "WEBHOOKS_UPDATE" => Self::WebhooksUpdate(WebhooksUpdate::deserialize(v)?),
-            other => {
-                return Err(ValueDeserializerError::UnknownVariant(
-                    other.to_owned(),
-                    &[],
-                ))
-            }
-        })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_guild() {
-        let broken_guild = r#"{
-  "d": {
-    "afk_channel_id": "1337",
-    "afk_timeout": 300,
-    "application_id": null,
-    "banner": null,
-    "default_message_notifications": 0,
-    "description": null,
-    "discovery_splash": null,
-    "embed_channel_id": null,
-    "embed_enabled": false,
-    "emojis": [
-      {
-        "animated": false,
-        "available": true,
-        "id": "1338",
-        "managed": false,
-        "name": "goodboi",
-        "require_colons": true,
-        "roles": []
-      }
-    ],
-    "explicit_content_filter": 0,
-    "features": [
-      "INVITE_SPLASH",
-      "ANIMATED_ICON"
-    ],
-    "guild_id": "1339",
-    "icon": "foobar",
-    "id": "13310",
-    "max_members": 250000,
-    "max_presences": null,
-    "mfa_level": 0,
-    "name": "FooBaz",
-    "owner_id": "13311",
-    "preferred_locale": "en-US",
-    "premium_subscription_count": 4,
-    "premium_tier": 1,
-    "region": "eu-central",
-    "roles": [
-      {
-        "color": 0,
-        "hoist": false,
-        "id": "13312",
-        "managed": false,
-        "mentionable": false,
-        "name": "@everyone",
-        "permissions": 104193601,
-        "position": 0
-      }
-    ],
-    "rules_channel_id": null,
-    "splash": "barbaz",
-    "system_channel_flags": 0,
-    "system_channel_id": "13313",
-    "vanity_url_code": null,
-    "verification_level": 0,
-    "widget_channel_id": null,
-    "widget_enabled": false
-  },
-  "op": 0,
-  "s": 42,
-  "t": "GUILD_UPDATE"
-}"#;
-
-        serde_json::from_str::<GatewayEvent>(broken_guild).unwrap();
-    }
-
-    #[test]
-    fn test_guild_2() {
-        let broken_guild = r#"{
-  "d": {
-    "afk_channel_id": null,
-    "afk_timeout": 300,
-    "application_id": null,
-    "banner": null,
-    "default_message_notifications": 0,
-    "description": null,
-    "discovery_splash": null,
-    "embed_channel_id": null,
-    "embed_enabled": true,
-    "emojis": [
-      {
-        "animated": false,
-        "available": true,
-        "id": "42",
-        "managed": false,
-        "name": "emmet",
-        "require_colons": true,
-        "roles": []
-      }
-    ],
-    "explicit_content_filter": 2,
-    "features": [],
-    "guild_id": "43",
-    "icon": "44",
-    "id": "45",
-    "max_members": 250000,
-    "max_presences": null,
-    "mfa_level": 0,
-    "name": "FooBar",
-    "owner_id": "46",
-    "preferred_locale": "en-US",
-    "premium_subscription_count": null,
-    "premium_tier": 0,
-    "region": "us-central",
-    "roles": [
-      {
-        "color": 0,
-        "hoist": false,
-        "id": "47",
-        "managed": false,
-        "mentionable": false,
-        "name": "@everyone",
-        "permissions": 104324673,
-        "position": 0
-      }
-    ],
-    "rules_channel_id": null,
-    "splash": null,
-    "system_channel_flags": 0,
-    "system_channel_id": "48",
-    "vanity_url_code": null,
-    "verification_level": 4,
-    "widget_channel_id": null,
-    "widget_enabled": true
-  },
-  "op": 0,
-  "s": 1190911,
-  "t": "GUILD_UPDATE"
-}"#;
-        serde_json::from_str::<GatewayEvent>(broken_guild).unwrap();
+        flags
     }
 }
