+++ gateway/src/shard/processor/emit.rs
@@ -1,12 +1,14 @@
-use super::super::event::{Event, EventType, Payload};
-use crate::listener::{Listener, Listeners};
+use crate::{
+    listener::{Listener, Listeners},
+    EventTypeFlags,
+};
 #[allow(unused_imports)]
 use log::{debug, info, trace, warn};
-//use tokio_executor::{DefaultExecutor, Executor};
+use twilight_model::gateway::event::{shard::Payload, Event};
 
 pub async fn bytes(listeners: Listeners<Event>, bytes: &[u8]) {
     for listener in listeners.all().lock().await.values() {
-        if listener.events.contains(EventType::SHARD_PAYLOAD) {
+        if listener.events.contains(EventTypeFlags::SHARD_PAYLOAD) {
             let event = Event::ShardPayload(Payload {
                 bytes: bytes.to_owned(),
             });
@@ -41,7 +43,7 @@ async fn _event(listeners: Listeners<Event>, event: Event) {
             break;
         }
 
-        let event_type = event.event_type();
+        let event_type = EventTypeFlags::from(event.kind());
 
         if !listener.events.contains(event_type) {
             trace!(
@@ -78,7 +80,7 @@ async fn _event(listeners: Listeners<Event>, event: Event) {
 /// If the receiver dropped, return `false` so we know to remove it.
 /// These are unbounded channels, so we know it's not because it's full.
 fn _emit_to_listener(id: u64, listener: &Listener<Event>, event: Event) -> bool {
-    let event_type = event.event_type();
+    let event_type = EventTypeFlags::from(event.kind());
 
     if !listener.events.contains(event_type) {
         trace!(
