+++ model/src/gateway/event/kind.rs
@@ -1,0 +1,54 @@
+/// The type of an event.
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+pub enum EventType {
+    BanAdd,
+    BanRemove,
+    ChannelCreate,
+    ChannelDelete,
+    ChannelPinsUpdate,
+    ChannelUpdate,
+    GatewayHeartbeat,
+    GatewayHeartbeatAck,
+    GatewayHello,
+    GatewayInvalidateSession,
+    GatewayReconnect,
+    GuildCreate,
+    GuildDelete,
+    GuildEmojisUpdate,
+    GuildIntegrationsUpdate,
+    GuildUpdate,
+    InviteCreate,
+    InviteDelete,
+    MemberAdd,
+    MemberRemove,
+    MemberUpdate,
+    MemberChunk,
+    MessageCreate,
+    MessageDelete,
+    MessageDeleteBulk,
+    MessageUpdate,
+    PresenceUpdate,
+    PresencesReplace,
+    ReactionAdd,
+    ReactionRemove,
+    ReactionRemoveAll,
+    ReactionRemoveEmoji,
+    Ready,
+    Resumed,
+    RoleCreate,
+    RoleDelete,
+    RoleUpdate,
+    ShardConnected,
+    ShardConnecting,
+    ShardDisconnected,
+    ShardIdentifying,
+    ShardReconnecting,
+    ShardPayload,
+    ShardResuming,
+    TypingStart,
+    UnavailableGuild,
+    UserUpdate,
+    VoiceServerUpdate,
+    VoiceStateUpdate,
+    WebhooksUpdate,
+}
